Update README.md
The teletype scroll function was not too difficult and there are many examples online. I had a project with an artist that required each word of the scrolling message to be colored differently than the previous. The colors also needed to change on each page load, and should only use complementary colors.

The central issue to overcome was that the text passed to the scroller JS function can contain no tags (e.g. <span style="color:#000BAA">Hello</span>. See the comments in the code to see how this was accomplished!